requirements.txt
anthropic>=0.42.0
python-dotenv>=1.0.0
pydantic>=2.0
httpx>=0.27.0
rich>=13.0
tavily-python>=0.5.0
